Add optional zbus implementation
