+++ Cargo.toml
@@ -26,8 +26,16 @@ dispatch = "0.2.0"
 objc = "0.2.7"
 
 [target.'cfg(all(unix, not(target_os = "macos")))'.dependencies]
-dbus = "0.9.5"
-dbus-crossroads = "0.5.0"
+dbus = { version = "0.9.5", optional = true }
+dbus-crossroads = { version = "0.5.0", optional = true }
+zbus = { version = "3.9", optional = true }
+zvariant = { version = "3.10", optional = true }
+pollster = { version = "0.3", optional = true }
+
+[features]
+default = ["use_dbus"]
+use_dbus = ["dbus", "dbus-crossroads"]
+use_zbus = ["zbus", "zvariant", "pollster"]
 
 [dev-dependencies]
 winit = "0.27.0"
